Connect monitor to pages (#78)
* feat: add monitors to pages

* feat: add monitors to pages

* feat: add monitors to pages

* feat: add monitors to pages

* chore: update migration

+++ packages/api/src/router/monitor.ts
@@ -2,9 +2,9 @@ import { z } from "zod";
 
 import { eq } from "@openstatus/db";
 import {
+  allMonitorsSchema,
   insertMonitorSchema,
   monitor,
-  selectMonitorSchema,
   user,
   usersToWorkspaces,
 } from "@openstatus/db/src/schema";
@@ -30,7 +30,7 @@ export const monitorRouter = createTRPCRouter({
         .get();
 
       // the user don't have access to this workspace
-      if (!result.users_to_workspaces) return;
+      if (!result || !result.users_to_workspaces) return;
 
       await opts.ctx.db.insert(monitor).values(opts.input).returning().get();
     }),
@@ -56,7 +56,7 @@ export const monitorRouter = createTRPCRouter({
         .innerJoin(currentUser, eq(usersToWorkspaces.userId, currentUser.id))
         .get();
 
-      if (!result.users_to_workspaces) return;
+      if (!result || !result.users_to_workspaces) return;
 
       return mon;
     }),
@@ -90,7 +90,7 @@ export const monitorRouter = createTRPCRouter({
         .innerJoin(currentUser, eq(usersToWorkspaces.userId, currentUser.id))
         .get();
 
-      if (!result.users_to_workspaces) return;
+      if (!result || !result.users_to_workspaces) return;
 
       opts.ctx.db
         .update(monitor)
@@ -120,7 +120,7 @@ export const monitorRouter = createTRPCRouter({
         .innerJoin(currentUser, eq(usersToWorkspaces.userId, currentUser.id))
         .get();
 
-      if (!result.users_to_workspaces) return;
+      if (!result || !result.users_to_workspaces) return;
 
       opts.ctx.db
         .update(monitor)
@@ -186,7 +186,7 @@ export const monitorRouter = createTRPCRouter({
         .innerJoin(currentUser, eq(usersToWorkspaces.userId, currentUser.id))
         .get();
 
-      if (!result.users_to_workspaces) return;
+      if (!result || !result.users_to_workspaces) return;
 
       opts.ctx.db.delete(monitor).where(eq(monitor.id, opts.input.monitorId));
     }),
@@ -206,17 +206,16 @@ export const monitorRouter = createTRPCRouter({
         )
         .innerJoin(currentUser, eq(usersToWorkspaces.userId, currentUser.id))
         .get();
-
       // the user don't have access to this workspace
-      if (!result.users_to_workspaces) return;
+      if (!result || !result.users_to_workspaces) return;
 
       const monitors = await opts.ctx.db
         .select()
         .from(monitor)
         .where(eq(monitor.workspaceId, opts.input.workspaceId))
         .all();
-      const selectMonitorsArray = selectMonitorSchema.array();
+      // const selectMonitorsArray = selectMonitorSchema.array();
 
-      return selectMonitorsArray.parse(monitors);
+      return allMonitorsSchema.parse(monitors);
     }),
 });
